requirements.txt
python-gdal
numpy
lxml
shapely
pyproj
cached_property
zipfile2
